some detail info

+++ app/src/main/java/com/ozontech/homework2/domain/mappers/ProductMapper.kt
@@ -7,7 +7,7 @@ import com.ozontech.homework2.presentation.viewObjects.ProductVO
 fun ProductDO.toVO(): ProductVO = ProductVO(
     guid,
     name,
-    price,
+    price.toInt(),
     description,
     rating.toFloat(),
     isFavorite,
